Move the SimpleFormatter parser tests to their own file
