+++ Tests/FoundationInternationalizationTests/Formatting/ListFormatStyleTests.swift
@@ -442,44 +442,4 @@ private struct ListFormatStyleTests {
         #expect([arabicIndic, arabicIndic].formatted(s) == "\(isolate(arabicIndic)) and \(isolate(arabicIndic))")
         #expect([persian, persian].formatted(s) == "\(persian) and \(persian)")
     }
-
-    // MARK: - Internal pattern parser
-
-    #if !FOUNDATION_LIST_FORMAT_ICU
-    /// `SimpleFormatter` decomposes patterns into prefix / connector / suffix at
-    /// formatter init. The format tests cover this indirectly; these cases pin
-    /// the parser's contract directly.
-    @Test func parseCanonicalPattern() {
-        let p = SimpleFormatter("{0}, {1}")
-        #expect(p?.prefix == "")
-        #expect(p?.connector == ", ")
-        #expect(p?.suffix == "")
-        #expect(p?.connectorStartsWithSpace == false)
-        #expect(p?.connectorEndsWithSpace == true)
-    }
-
-    @Test func parseSpanishPattern() {
-        let p = SimpleFormatter("{0} y {1}")
-        #expect(p?.connector == " y ")
-        #expect(p?.connectorStartsWithSpace == true)
-        #expect(p?.connectorEndsWithSpace == true)
-    }
-
-    @Test func parseWithPrefixAndSuffix() {
-        let p = SimpleFormatter("foo{0}…{1}bar")
-        #expect(p?.prefix == "foo")
-        #expect(p?.connector == "…")
-        #expect(p?.suffix == "bar")
-        #expect(p?.connectorStartsWithSpace == false)
-        #expect(p?.connectorEndsWithSpace == false)
-    }
-
-    @Test func parseRejectsMalformed() {
-        #expect(SimpleFormatter("garbage") == nil)
-        #expect(SimpleFormatter("{0}") == nil)            // missing {1}
-        #expect(SimpleFormatter("{1}, {0}") == nil)       // wrong order
-        #expect(SimpleFormatter("{0, {1}") == nil)        // unclosed brace
-        #expect(SimpleFormatter("") == nil)
-    }
-    #endif
 }
